refactor: delete deck refactoring
