+++ src/components/ui/layout-components/tables/deck-list-table/deck-list-table.tsx
@@ -115,13 +115,15 @@ export const DeckListTable = ({ decks, user }: DecksListTableProps) => {
         </TableRow>
       </TableHeader>
       <TableBody>{tableRows}</TableBody>
-      <DeckDialogForm
-        action={DIALOG_ACTION.UPDATE}
-        deck={deckData}
-        key={deckData?.id}
-        onOpenChange={setShowEditDeckDialog}
-        open={showEditDeckDialog}
-      />
+      {showEditDeckDialog && (
+        <DeckDialogForm
+          action={DIALOG_ACTION.UPDATE}
+          deck={deckData}
+          key={deckData?.id}
+          onOpenChange={setShowEditDeckDialog}
+          open={showEditDeckDialog}
+        />
+      )}
       <DeleteDialogForm
         entity={DIALOG_ENTITY.DECK}
         entityId={deckId}
