+++ src/shared/hooks/index.ts
@@ -1,9 +1,12 @@
 export * from './use-auth-context'
+export * from './use-card-dialog-form-data'
 export * from './use-card-page-data'
 export * from './use-current-screen-width'
+export * from './use-deck-dialog-form-data'
 export * from './use-deck-list-page-data'
 export * from './use-deck-list-table-data'
 export * from './use-deck-page-data'
 export * from './use-deck-table-data'
+export * from './use-delete-dialog-form-data'
 export * from './use-profile-data'
 export * from './use-search-param-updater'
